test new login redirect
