+++ frontend/package.json
@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "@bcgov/bc-sans": "2.1.0",
-    "@bcgov/citz-imb-sso-react": "1.0.0-alpha2",
+    "@bcgov/citz-imb-sso-react": "file:./sso-react-1.0.0-alpha3.tgz",
     "@bcgov/design-system-react-components": "0.0.6",
     "@vitejs/plugin-react": "4.2.1",
     "react": "18.2.0",
