+++ frontend/src/main.tsx
@@ -1,14 +1,14 @@
+import { SSOProvider } from '@bcgov/citz-imb-sso-react';
 import 'css/base.css';
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { SSOProvider } from '@bcgov/citz-imb-sso-react';
-import AppRouter from 'AppRouter';
+import App from './App';
 
 const root = createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <React.StrictMode>
     <SSOProvider>
-      <AppRouter />
+      <App />
     </SSOProvider>
   </React.StrictMode>,
 );
